+++ semialign/src/Data/Semialign/Indexed.hs
@@ -1,0 +1,9 @@
+{-# LANGUAGE Safe #-}
+-- | Zipping and aligning of indexed functors.
+module Data.Semialign.Indexed (
+    SemialignWithIndex (..),
+    ZipWithIndex (..),
+    RepeatWithIndex (..),
+) where
+
+import Data.Semialign.Internal